+++ src/features/user/auth/index.ts
@@ -1,7 +1,2 @@
-// Hooks
 export { useAuth } from "./hooks/useAuth";
 export { useAuthHandlers } from "./hooks/useAuthHandlers";
-export { useUserSessions } from "./hooks/useUserSessions";
-
-// Utils
-export { isCurrentSession } from "./utils/session";